# Implement `fast_float` for `String#to_f`