--- a/spec/std/string/fast_float_spec.cr
+++ b/spec/std/string/fast_float_spec.cr
@@ -1,0 +1,5 @@
+require "spec"
+
+describe String do
+  # TODO
+end