--- a/spec/std/string/fast_float_spec.cr
+++ b/spec/std/string/fast_float_spec.cr
@@ -1,5 +1,0 @@
-require "spec"
-
-describe String do
-  # TODO
-end