--- a/src/lib_c/x86_64-windows-msvc/c/stdlib.cr
+++ b/src/lib_c/x86_64-windows-msvc/c/stdlib.cr
@@ -11,13 +11,13 @@ lib LibC
   fun free(ptr : Void*) : Void
   fun malloc(size : SizeT) : Void*
   fun realloc(ptr : Void*, size : SizeT) : Void*
-  fun strtof(nptr : Char*, endptr : Char**) : Float
-  fun strtod(nptr : Char*, endptr : Char**) : Double
 
   alias InvalidParameterHandler = WCHAR*, WCHAR*, WCHAR*, UInt, UIntPtrT ->
   fun _set_invalid_parameter_handler(pNew : InvalidParameterHandler) : InvalidParameterHandler
 
   # unused
+  fun strtof(nptr : Char*, endptr : Char**) : Float
+  fun strtod(nptr : Char*, endptr : Char**) : Double
   fun atof(nptr : Char*) : Double
   fun div(numer : Int, denom : Int) : DivT
   fun putenv(string : Char*) : Int