… spec/manual/string_to_f_supplemental_spec.cr
… spec/support/number.cr
… src/float/fast_float/decimal_to_binary.cr
… src/float/fast_float/fast_table.cr
… src/float/fast_float/float_common.cr