@kostya 2024-11-19
I also added shard for this: https://github.com/kostya/fast_to_f